Improved base Docker image (#28)
Partially addresses #14, #12, #11. A Dockerfile is developed from scratch to address size & dependency issues. 3D inline rendering, properly configured JupyterLab, and required dependencies and tools are some of the features of this Dockerfile over the current one. In addition, extra care is taken to make the container size as small as possible.